Update range in removeTokenIfPresent
When the label is preloaded with text from a xib or storyboard, the truncatedRange is never set. This would cause a crash when this was hit. To resolve this, I used the local truncationRange. The truncationRange variable should always be correct as it is built based off the class variable textStorage. 

Tested on my application and it works with preloaded strings, strings that don't require truncation, and strings that require truncation.